requirements.txt
temporalio==1.21.1
httpx==0.28.1
python-dotenv==1.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0

